Update CC0
I'm pretty sure this is I don't know. I was trying to make this a public domain but like I am not that smart for this sort of thing like I just find myself staring at this website like and there are just so many buttons. But anyway, hopefully this fixes what was happening with the you know whatever you get it. Geez