Update CC0
So anyways this is the chatgpt video game I wanted to make but like I guess I just got so much s*** I want to do. Like obviously I just don't have time for this and really this is really cool. I actually want to play this game so somebody could make this like that'd be really cool